requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==1.16.0
idna==2.10
mysql-connector-python==8.0.23
numpy==1.18.5
pandas==0.24.2
protobuf==3.15.8
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
